// src/app/interfaces/book-response.ts
export interface BookDetails {
  error:     string;
  title:     string;
  subtitle:  string;
  authors:   string;
  publisher: string;
  language:  string;
  isbn10:    string;
  isbn13:    string;
  pages:     string;
  year:      string;
  rating:    string;
  desc:      string;
  price:     string;
  image:     string;
  url:       string;
}